Rename `channel_view.py` to `channel.py`
