+++ src/pathfinding_service/model/__init__.py
@@ -1,4 +1,4 @@
-from .channel_view import ChannelView
+from .channel import ChannelView
 from .feedback import FeedbackToken
 from .iou import IOU
 from .token_network import TokenNetwork
