+++ src/pathfinding_service/model/channel.py
