+++ tests/pathfinding/test_graphs.py
@@ -8,7 +8,7 @@ from eth_utils import decode_hex, to_canonical_address, to_checksum_address
 
 from pathfinding_service.constants import DIVERSITY_PEN_DEFAULT
 from pathfinding_service.model import ChannelView, TokenNetwork
-from pathfinding_service.model.channel_view import Channel
+from pathfinding_service.model.channel import Channel
 from raiden.network.transport.matrix import AddressReachability
 from raiden.utils.typing import (
     Address,
